Rail Shooter game.
This is rail shooter game in  which enemies will fly on the specified path (called rail) the player have to shoot them or dodge them.
Player can move verticle and horizontal and can shoot durig the movement.

The core game loop is to fly as far as possible without dying in order to achieve the highest possible score.The game will restart after the player dies.